+++ xrpl/asyncio/transaction/main.py
@@ -30,9 +30,8 @@ _LEDGER_OFFSET: Final[int] = 20
 # More context: https://github.com/XRPLF/rippled/pull/4370
 _RESTRICTED_NETWORKS = 1024
 _REQUIRED_NETWORKID_VERSION = "1.11.0"
-_HOOKS_TESTNET_ID = 21338
 # TODO: make this dynamic based on the current ledger fee
-_ACCOUNT_DELETE_FEE: Final[int] = int(xrp_to_drops(2))
+_OWNER_RESERVE_FEE: Final[int] = int(xrp_to_drops(2))
 
 
 async def sign_and_submit(
@@ -283,14 +282,9 @@ def _tx_needs_networkID(client: Client) -> bool:
         bool: whether the transactions required network ID to be valid
     """
     if client.network_id and client.network_id > _RESTRICTED_NETWORKS:
-        # TODO: remove the buildVersion logic when 1.11.0 is out and widely used.
-        # Issue: https://github.com/XRPLF/xrpl-py/issues/595
-        if (
-            client.build_version
-            and _is_not_later_rippled_version(
-                _REQUIRED_NETWORKID_VERSION, client.build_version
-            )
-        ) or client.network_id == _HOOKS_TESTNET_ID:
+        if client.build_version and _is_not_later_rippled_version(
+            _REQUIRED_NETWORKID_VERSION, client.build_version
+        ):
             return True
     return False
 
@@ -467,11 +461,14 @@ async def _calculate_fee_per_transaction_type(
             base_fee = math.ceil(net_fee * (33 + (len(fulfillment_bytes) / 16)))
 
     # AccountDelete Transaction
-    if transaction.transaction_type == TransactionType.ACCOUNT_DELETE:
+    if transaction.transaction_type in (
+        TransactionType.ACCOUNT_DELETE,
+        TransactionType.AMM_CREATE,
+    ):
         if client is None:
-            base_fee = _ACCOUNT_DELETE_FEE
+            base_fee = _OWNER_RESERVE_FEE
         else:
-            base_fee = await _fetch_account_delete_fee(client)
+            base_fee = await _fetch_owner_reserve_fee(client)
 
     # Multi-signed Transaction
     # 10 drops × (1 + Number of Signatures Provided)
@@ -481,7 +478,7 @@ async def _calculate_fee_per_transaction_type(
     return str(math.ceil(base_fee))
 
 
-async def _fetch_account_delete_fee(client: Client) -> int:
+async def _fetch_owner_reserve_fee(client: Client) -> int:
     server_state = await client._request_impl(ServerState())
     fee = server_state.result["state"]["validated_ledger"]["reserve_inc"]
     return int(fee)