+++ xrpl/models/requests/ledger_entry.py
@@ -15,6 +15,7 @@ from xrpl.models.base_model import BaseModel
 from xrpl.models.requests.request import LookupByLedgerRequest, Request, RequestMethod
 from xrpl.models.required import REQUIRED
 from xrpl.models.utils import require_kwargs_on_init
+from xrpl.models.xchain_bridge import XChainBridge
 
 
 class LedgerEntryType(str, Enum):
@@ -22,6 +23,7 @@ class LedgerEntryType(str, Enum):
 
     ACCOUNT = "account"
     AMENDMENTS = "amendments"
+    AMM = "amm"
     CHECK = "check"
     DEPOSIT_PREAUTH = "deposit_preauth"
     DIRECTORY = "directory"
@@ -132,7 +134,7 @@ class Offer(BaseModel):
 @require_kwargs_on_init
 @dataclass(frozen=True)
 class RippleState(BaseModel):
-    """Required fields for requesting a RippleState."""
+    """Required fields for requesting a RippleState if not querying by object ID."""
 
     accounts: List[str] = REQUIRED  # type: ignore
     """
@@ -152,10 +154,7 @@ class RippleState(BaseModel):
 @require_kwargs_on_init
 @dataclass(frozen=True)
 class Ticket(BaseModel):
-    """
-    Required fields for requesting a Ticket, if not querying by
-    object ID.
-    """
+    """Required fields for requesting a Ticket if not querying by object ID."""
 
     owner: str = REQUIRED  # type: ignore
     """
@@ -167,6 +166,37 @@ class Ticket(BaseModel):
     ticket_sequence: int = REQUIRED  # type: ignore
     """
     This field is required.
+
+    :meta hide-value:
+    """
+
+
+@require_kwargs_on_init
+@dataclass(frozen=True)
+class XChainClaimID(XChainBridge):
+    """Required fields for requesting an XChainClaimID if not querying by object ID."""
+
+    xchain_claim_id: Union[int, str] = REQUIRED  # type: ignore
+    """
+    The `XChainClaimID` associated with a cross-chain transfer, which was created in an
+    `XChainCreateClaimID` transaction. This field is required.
+
+    :meta hide-value:
+    """
+
+
+@require_kwargs_on_init
+@dataclass(frozen=True)
+class XChainCreateAccountClaimID(XChainBridge):
+    """
+    Required fields for requesting an XChainCreateAccountClaimID if not querying by
+    object ID.
+    """
+
+    xchain_create_account_claim_id: Union[int, str] = REQUIRED  # type: ignore
+    """
+    The `XChainCreateAccountClaimID` associated with a cross-chain account create. This
+    field is required.
 
     :meta hide-value:
     """
@@ -194,6 +224,13 @@ class LedgerEntry(Request, LookupByLedgerRequest):
     payment_channel: Optional[str] = None
     ripple_state: Optional[RippleState] = None
     ticket: Optional[Union[str, Ticket]] = None
+    bridge_account: Optional[str] = None
+    bridge: Optional[XChainBridge] = None
+    xchain_claim_id: Optional[Union[str, XChainClaimID]] = None
+    xchain_create_account_claim_id: Optional[
+        Union[str, XChainCreateAccountClaimID]
+    ] = None
+
     binary: bool = False
     nft_page: Optional[str] = None
     """Must be the object ID of the NFToken page, as hexadecimal"""
@@ -205,17 +242,28 @@ class LedgerEntry(Request, LookupByLedgerRequest):
             for param in [
                 self.index,
                 self.account_root,
-                self.directory,
-                self.offer,
-                self.ripple_state,
                 self.check,
-                self.escrow,
-                self.payment_channel,
                 self.deposit_preauth,
+                self.directory,
+                self.escrow,
+                self.offer,
+                self.payment_channel,
+                self.ripple_state,
                 self.ticket,
+                self.xchain_claim_id,
+                self.xchain_create_account_claim_id,
             ]
             if param is not None
         ]
-        if len(query_params) != 1:
+        if (
+            len(query_params) != 1
+            and self.bridge is None
+            and self.bridge_account is None
+        ):
             errors["LedgerEntry"] = "Must choose exactly one data to query"
+
+        if (self.bridge is None) != (self.bridge_account is None):
+            # assert that you either have both of these or neither
+            errors["Bridge"] = "Must include both `bridge` and `bridge_account`."
+
         return errors