+++ xrpl/models/transactions/__init__.py
@@ -10,9 +10,24 @@ from xrpl.models.transactions.account_set import (
     AccountSetFlag,
     AccountSetFlagInterface,
 )
+from xrpl.models.transactions.amm_bid import AMMBid, AuthAccount
+from xrpl.models.transactions.amm_create import AMMCreate
+from xrpl.models.transactions.amm_delete import AMMDelete
+from xrpl.models.transactions.amm_deposit import (
+    AMMDeposit,
+    AMMDepositFlag,
+    AMMDepositFlagInterface,
+)
+from xrpl.models.transactions.amm_vote import AMMVote
+from xrpl.models.transactions.amm_withdraw import (
+    AMMWithdraw,
+    AMMWithdrawFlag,
+    AMMWithdrawFlagInterface,
+)
 from xrpl.models.transactions.check_cancel import CheckCancel
 from xrpl.models.transactions.check_cash import CheckCash
 from xrpl.models.transactions.check_create import CheckCreate
+from xrpl.models.transactions.clawback import Clawback
 from xrpl.models.transactions.deposit_preauth import DepositPreauth
 from xrpl.models.transactions.escrow_cancel import EscrowCancel
 from xrpl.models.transactions.escrow_create import EscrowCreate
@@ -54,6 +69,24 @@ from xrpl.models.transactions.trust_set import (
     TrustSetFlag,
     TrustSetFlagInterface,
 )
+from xrpl.models.transactions.xchain_account_create_commit import (
+    XChainAccountCreateCommit,
+)
+from xrpl.models.transactions.xchain_add_account_create_attestation import (
+    XChainAddAccountCreateAttestation,
+)
+from xrpl.models.transactions.xchain_add_claim_attestation import (
+    XChainAddClaimAttestation,
+)
+from xrpl.models.transactions.xchain_claim import XChainClaim
+from xrpl.models.transactions.xchain_commit import XChainCommit
+from xrpl.models.transactions.xchain_create_bridge import XChainCreateBridge
+from xrpl.models.transactions.xchain_create_claim_id import XChainCreateClaimID
+from xrpl.models.transactions.xchain_modify_bridge import (
+    XChainModifyBridge,
+    XChainModifyBridgeFlag,
+    XChainModifyBridgeFlagInterface,
+)
 
 __all__ = [
     "AccountDelete",
@@ -61,9 +94,21 @@ __all__ = [
     "AccountSetAsfFlag",
     "AccountSetFlag",
     "AccountSetFlagInterface",
+    "AMMBid",
+    "AMMCreate",
+    "AMMDelete",
+    "AMMDeposit",
+    "AMMDepositFlag",
+    "AMMDepositFlagInterface",
+    "AMMVote",
+    "AMMWithdraw",
+    "AMMWithdrawFlag",
+    "AMMWithdrawFlagInterface",
+    "AuthAccount",
     "CheckCancel",
     "CheckCash",
     "CheckCreate",
+    "Clawback",
     "DepositPreauth",
     "EscrowCancel",
     "EscrowCreate",
@@ -100,4 +145,14 @@ __all__ = [
     "TrustSet",
     "TrustSetFlag",
     "TrustSetFlagInterface",
+    "XChainAccountCreateCommit",
+    "XChainAddAccountCreateAttestation",
+    "XChainAddClaimAttestation",
+    "XChainClaim",
+    "XChainCommit",
+    "XChainCreateBridge",
+    "XChainCreateClaimID",
+    "XChainModifyBridge",
+    "XChainModifyBridgeFlag",
+    "XChainModifyBridgeFlagInterface",
 ]