+++ xrpl/models/transactions/transaction.py
@@ -9,7 +9,7 @@ from typing_extensions import Final
 
 from xrpl.core.binarycodec import decode, encode
 from xrpl.models.amounts import IssuedCurrencyAmount
-from xrpl.models.base_model import BaseModel
+from xrpl.models.base_model import ABBREVIATIONS, BaseModel
 from xrpl.models.exceptions import XRPLModelException
 from xrpl.models.flags import check_false_flag_definition, interface_to_flag_list
 from xrpl.models.nested_model import NestedModel
@@ -20,14 +20,6 @@ from xrpl.models.types import XRPL_VALUE_TYPE
 from xrpl.models.utils import require_kwargs_on_init
 
 _TRANSACTION_HASH_PREFIX: Final[int] = 0x54584E00
-# This is used to make exceptions when converting dictionary keys to xrpl JSON
-# keys. We snake case keys, but some keys are abbreviations.
-_ABBREVIATIONS: Final[Dict[str, str]] = {
-    "unl": "UNL",
-    "id": "ID",
-    "uri": "URI",
-    "nftoken": "NFToken",
-}
 
 
 def transaction_json_to_binary_codec_form(
@@ -56,11 +48,11 @@ def _key_to_tx_json(key: str) -> str:
         1. 'transaction_type' becomes 'TransactionType'
         2. 'URI' becomes 'uri'
 
-    Known abbreviations (example 2 above) need to be enumerated in _ABBREVIATIONS.
+    Known abbreviations (example 2 above) need to be enumerated in ABBREVIATIONS.
     """
     return "".join(
         [
-            _ABBREVIATIONS[word] if word in _ABBREVIATIONS else word.capitalize()
+            ABBREVIATIONS[word] if word in ABBREVIATIONS else word.capitalize()
             for word in key.split("_")
         ]
     )
@@ -69,7 +61,9 @@ def _key_to_tx_json(key: str) -> str:
 def _value_to_tx_json(value: XRPL_VALUE_TYPE) -> XRPL_VALUE_TYPE:
     # IssuedCurrencyAmount and PathStep are special cases and should not be snake cased
     # and only contain primitive members
-    if IssuedCurrencyAmount.is_dict_of_model(value) or PathStep.is_dict_of_model(value):
+    if isinstance(value, list) and all(PathStep.is_dict_of_model(v) for v in value):
+        return value
+    if IssuedCurrencyAmount.is_dict_of_model(value):
         return value
     if isinstance(value, dict):
         return transaction_json_to_binary_codec_form(value)
@@ -259,6 +253,9 @@ class Transaction(BaseModel):
     """The network id of the transaction."""
 
     def _get_errors(self: Transaction) -> Dict[str, str]:
+        # import must be here to avoid circular dependencies
+        from xrpl.wallet.main import Wallet
+
         errors = super()._get_errors()
         if self.ticket_sequence is not None and (
             (self.sequence is not None and self.sequence != 0)
@@ -268,6 +265,10 @@ class Transaction(BaseModel):
                 "Transaction"
             ] = """If ticket_sequence is provided,
             account_txn_id must be None and sequence must be None or 0"""
+
+        if isinstance(self.account, Wallet):
+            errors["account"] = "Must pass in `wallet.address`, not `wallet`."
+
         return errors
 
     def to_dict(self: Transaction) -> Dict[str, Any]: