+++ xrpl/utils/get_nftoken_id.py
@@ -1,5 +1,5 @@
 """Utils to get an NFTokenID from metadata"""
-from typing import Callable, List, TypeVar
+from typing import Callable, List, TypeVar, Union
 
 from xrpl.models.transactions.metadata import (
     NFTokenMetadata,
@@ -49,7 +49,7 @@ def get_nftoken_ids_from_nftokens(nftokens: List[NFTokenMetadata]) -> List[str]:
     ]
 
 
-def get_nftoken_id(meta: TransactionMetadata) -> str:
+def get_nftoken_id(meta: TransactionMetadata) -> Union[str, None]:
     """
     Gets the NFTokenID for an NFT recently minted with NFTokenMint.
 
@@ -57,7 +57,8 @@ def get_nftoken_id(meta: TransactionMetadata) -> str:
         meta: Metadata from the response to submitting an NFTokenMint transaction.
 
     Returns:
-        The newly minted NFToken's NFTokenID.
+        The newly minted NFToken's NFTokenID. None if there is no NFTokenID that was
+        minted.
 
     Raises:
         TypeError: if given something other than metadata (e.g. the full
@@ -96,6 +97,9 @@ def get_nftoken_id(meta: TransactionMetadata) -> str:
             return False
 
     affected_nodes = [node for node in meta["AffectedNodes"] if has_nftoken_page(node)]
+
+    if len(affected_nodes) == 0:
+        return None
 
     def get_previous_nftokens(node: Node) -> List[NFTokenMetadata]:
         nftokens: List[NFTokenMetadata] = []